+++ examples/express-checkout/src/App.vue
@@ -179,11 +179,11 @@ onMounted(async () => {
         </a>
         <div class="px-5 pb-5">
           <a href="#">
-            <h5
+            <div
               class="text-xl font-semibold tracking-tight text-gray-900 dark:text-white"
             >
               {{ productName }}
-            </h5>
+            </div>
           </a>
           <div class="flex items-center justify-between">
             <span class="text-3xl font-bold text-gray-900 dark:text-white">{{